server/requirements.txt
fastapi
uvicorn[standard]
pandas
google-generativeai
python-dotenv
numpy
scikit-learn
httpx
aiohttp
